imx8mm_cpuid: add driver to report the CPU ID
This component reads the OCOTP_HW_OCOTP_TESTER0 ant OCOTP_HW_OCOTP_TESTER1
registers to get the CPUID.
See https://community.nxp.com/t5/i-MX-Processors/how-to-get-CPU-id-of-imx8mm/m-p/1317003

The output of this component is used by `uspclient` as the serial number
of the device.

There is a run script `run/cpuid` that is executed on the pipeline to
test the component.

Issue #1396